AIRA-35: Refactor - Rename Project from 'Compass' to 'Aira'
