+++ README.md
@@ -1,6 +1,6 @@
-# 🧭 Compass: The Autonomous Incident Response Agent
+# 🧭 Aira: The Autonomous Incident Response Agent
 
 
-**Compass is the open-source AI agent that acts as an autonomous Site Reliability Engineer (SRE). When an incident occurs, Compass automatically connects to your essential services—like PagerDuty, Datadog, GitHub, and AWS—to gather critical context. It then uses a Large Language Model (LLM) to analyze the correlated data and deliver a clear, probable root-cause hypothesis directly to your team, turning hours of stressful manual investigation into a minutes-long, automated process.**
+**Aira is the open-source AI agent that acts as an autonomous Site Reliability Engineer (SRE). When an incident occurs, Aira automatically connects to your essential services—like PagerDuty, Datadog, GitHub, and AWS—to gather critical context. It then uses a Large Language Model (LLM) to analyze the correlated data and deliver a clear, probable root-cause hypothesis directly to your team, turning hours of stressful manual investigation into a minutes-long, automated process.**
 
 ---
