+++ aira/connectors/alerting/__init__.py
@@ -1,7 +1,7 @@
-# compass/connectors/collaboration/__init__.py
+# aira/connectors/alerting/__init__.py
 
-from .slack import SlackConnector
+from .pagerduty import PagerDutyConnector
 
 # This line explicitly declares which names are part of this package's
 # public interface, silencing the "unused import" warning.
-__all__ = ["SlackConnector"]
+__all__ = ["PagerDutyConnector"]
