+++ aira/connectors/collaboration/__init__.py
@@ -1,7 +1,7 @@
-# compass/connectors/alerting/__init__.py
+# aira/connectors/collaboration/__init__.py
 
-from .pagerduty import PagerDutyConnector
+from .slack import SlackConnector
 
 # This line explicitly declares which names are part of this package's
 # public interface, silencing the "unused import" warning.
-__all__ = ["PagerDutyConnector"]
+__all__ = ["SlackConnector"]
