+++ aira/templates/config.template.yaml
@@ -1,8 +1,8 @@
-# 🧭 Compass Configuration File
+# 🧭 Aira Configuration File
 # --------------------------------------------------------------------
-# This file was generated by `compass init`.
+# This file was generated by `aira init`.
 # Secrets like API keys are loaded from your environment variables for security.
-# For local development, you can create a .env file in ~/.compass/
+# For local development, you can create a .env file in ~/.aira/
 # --------------------------------------------------------------------
 
 # --- LLM Provider Configuration ---
