+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-env_files = /Users/riansh/.compass/.env
+env_files = ~/.aira/.env
